fix(ATT-101): ensure translations do not cause unnecessary rerender
