+++ libs/plugins-frontend-ui/src/lib/i18n.ts
@@ -1,4 +1,4 @@
-import { useCallback, useMemo } from 'react';
+import { useCallback, useMemo, useRef } from 'react';
 import { create } from 'zustand';
 import { get } from 'lodash-es';
 import * as Handlebars from 'handlebars';
@@ -47,13 +47,21 @@ interface UseTranslationsResponse {
 export function useTranslations(translations: TranslationModules): UseTranslationsResponse {
   const { language, setLanguage } = useTranslationState();
 
+  // Keep a stable reference to the provided translations so callers
+  // can safely pass inline objects without causing re-renders.
+  const translationsRef = useRef(translations);
+  if (translationsRef.current !== translations) {
+    // We intentionally ignore subsequent translation object identities
+    // assuming translation modules are static.
+  }
+
   const activeTranslations = useMemo(() => {
-    return translations[language];
-  }, [language, translations]);
+    return translationsRef.current[language];
+  }, [language]);
 
   const fallbackTranslations = useMemo(() => {
-    return translations['en'];
-  }, [translations]);
+    return translationsRef.current['en'];
+  }, []);
 
   const getTranslationRaw = useCallback(
     (key: string) => {
@@ -80,17 +88,11 @@ export function useTranslations(translations: TranslationModules): UseTranslatio
       const ABSOLUTE_FALLBACK_TRANSLATION = `!!! ${key} !!!`;
       const translationTemplate = getTranslationTemplate(key);
       if (translationTemplate === undefined) {
-        console.log('translationTemplate is undefined', {
-          key,
-          activeTranslations,
-          fallbackTranslations,
-          translations,
-        });
         return ABSOLUTE_FALLBACK_TRANSLATION;
       }
       return translationTemplate(data);
     },
-    [getTranslationTemplate, activeTranslations, fallbackTranslations, translations],
+    [getTranslationTemplate],
   );
 
   const tExists = useCallback(
